potentiometer driver
Added a potentiometer driver for both linear and logarithmic tapers. Potentiometer experiences jitter, need to add a LPF to eliminate ADC noise.

+++ Poly3/Core/Inc/encoder.h
@@ -26,7 +26,7 @@ typedef enum {
 
 
 typedef struct {
-	int32_t index;
+	int8_t index;
     int32_t position;
     encoder_direction_t last_dir;
     uint8_t state;
